+++ picasso/server/helper.py
@@ -6,14 +6,15 @@ from sqlalchemy import create_engine
 import pandas as pd
 import streamlit as st
 import time
-
-
-def _db_filename():
-    home = os.path.expanduser("~")
-    return os.path.abspath(os.path.join(home, ".picasso", "app.db"))
+import subprocess
+import picasso.localize
+from picasso.localize import _db_filename
 
 
 def fetch_db():
+    """
+    Helper function to load the local database and return the files.
+    """
     try:
         engine = create_engine("sqlite:///" + _db_filename(), echo=False)
         df = pd.read_sql_table("files", con=engine)
@@ -26,6 +27,9 @@ def fetch_db():
 
 
 def fetch_watcher():
+    """
+    Helper function to load the local database and return running watchers.
+    """
     try:
         engine = create_engine("sqlite:///" + _db_filename(), echo=False)
         df = pd.read_sql_table("watcher", con=engine)
@@ -37,10 +41,10 @@ def fetch_watcher():
 
 def refresh(to_wait: int):
     """
-    Utility function that waits for a given amount and then restarts streamlit.
+    Utility function that waits for a given amount and then stops streamlit.
     """
     ref = st.empty()
     for i in range(to_wait):
         ref.write(f"Refreshing in {to_wait-i} s")
         time.sleep(1)
-    raise st.script_runner.RerunException(st.script_request_queue.RerunData(None))
+    st.stop()