+++ changelog.rst
@@ -1,9 +1,17 @@
 Changelog
 =========
 
-Last change: 23-MAR-2022 MTS
+Last change: 01-JUL-2022 MTS
 
-0.4.0
+0.4.2
+-----
+- Added more docstrings / documentation for Picasso Server
+- Import and export for handling IMS (Imaris) files
+- Fixed a bug where GPUFit was greyed out, added better installation instructions for GPUfit
+- More documentation
+- Added dockerfile
+
+0.4.1
 -----
 - Fixed a bug in installation